Define variable to only highlight current window
I believe before 1.61 only the current window was highlighted:
https://github.com/jcs-elpa/auto-highlight-symbol/commit/b3674dd553e49a52f9a0ebb2085446ac9ac7b7c4#diff-e7a9435d360fadbeb259c0d4db30503bbaf19ae209f140595f61ac60279d9ba6R942

This PR introduces a way to restore that ability, but preserves the 1.61
behavior as the default.

No pressure to adopt this, of course, if this isn't the direction you want your
package to go <3 I'm totally happy to maintain a fork for myself since I
understand my use-cases may be a bit different from your other users!