Define variable to not highlight upon window switch
I believe before 1.61 focus-switching would not trigger highlighting:
https://github.com/jcs-elpa/auto-highlight-symbol/commit/a79ce27346dc658379df0adf4e0e758d29437bc5

This PR introduces a way to restore that ability, but preserves the 1.61
behavior as the default.

No pressure to adopt this, of course, if this isn't the direction you want your
package to go <3 I'm totally happy to maintain a fork for myself since I
understand my use-cases may be a bit different from your other users!

As context, my use-case is to trigger highlighting only upon invocation of the
[`symbol-navigation-hydra`](https://github.com/bgwines/symbol-navigation-hydra).